% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid.R
\name{aa_index}
\alias{aa_index}
\title{return the amino acid index in this ORF}
\usage{
aa_index(x, start, end)
}
\arguments{
\item{x}{1-based nt index}

\item{start}{ORF start}

\item{end}{ORF end}
}
\value{
amino acid index
}
\description{
return the amino acid index in this ORF
}
\examples{
aa_index(5, 1, 12)
}
