% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_bio.R
\name{antigen_map}
\alias{antigen_map}
\title{build antigen map from sera titer data}
\usage{
antigen_map(
  data,
  sera_meta = NULL,
  ag_meta = NULL,
  value_lim = c(0, Inf),
  n_optim = 500,
  seed = NULL
)
}
\arguments{
\item{data}{titer tibble, colnames should be \verb{id, antigen1, antigen2, ...}}

\item{sera_meta}{metadata of sera, colnames should be \verb{id, arg1, arg2, ...}}

\item{ag_meta}{metadata of antigen, colnames should be \verb{id, arg1, arg2, ...}}

\item{value_lim}{titer value limitation, default as \code{c(0, Inf)}}

\item{n_optim}{optimization runs to perform}

\item{seed}{random seed}
}
\value{
tibble
}
\description{
build antigen map from sera titer data
}
