% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid.R
\name{codon_index}
\alias{codon_index}
\title{return the index in the correspond codon (1, 2 or 3)}
\usage{
codon_index(x, start, end)
}
\arguments{
\item{x}{1-based nt index}

\item{start}{ORF start}

\item{end}{ORF end}
}
\value{
1, 2 or 3
}
\description{
return the index in the correspond codon (1, 2 or 3)
}
\examples{
codon_index(5, 1, 12)
}
