% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid.R
\name{codon_range}
\alias{codon_range}
\title{return the codon range of the correspond codon}
\usage{
codon_range(x, start, end)
}
\arguments{
\item{x}{1-based nt index}

\item{start}{ORF start}

\item{end}{ORF end}
}
\value{
codon range string
}
\description{
return the codon range of the correspond codon
}
\examples{
codon_range(5, 1, 12)
}
