% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genogamesh-package.R
\docType{package}
\name{genogamesh-package}
\alias{genogamesh}
\alias{genogamesh-package}
\title{genogamesh: Process genomics data}
\description{
Process genomics data
}
\author{
\strong{Maintainer}: William Song \email{william_swl@163.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
