% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutstr.R
\docType{class}
\name{mutstr-class}
\alias{mutstr-class}
\title{a S4 class to operate mutation strings}
\description{
a S4 class to operate mutation strings
}
\section{Slots}{

\describe{
\item{\code{names}}{character. name of each item}

\item{\code{mstr}}{character. raw mutation strings}

\item{\code{sep}}{character. separation between two mutations}

\item{\code{mut}}{list. splited mutation list}
}}

