% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutstr.R
\name{mutstr}
\alias{mutstr}
\title{create a mutstr object}
\usage{
mutstr(mstr = "", names = NULL, sep = ",")
}
\arguments{
\item{mstr}{character vector of mutations}

\item{names}{item names}

\item{sep}{separation of each mutation}
}
\value{
mutstr
}
\description{
create a mutstr object
}
\examples{

raw_mut_string <- c(
  variant1 = "T10I,D20N,Q30E,A40T,P50L,G60R",
  variant2 = "T10I,D20-,Q30E,A40T,P50L,G60R,S80R",
  variant3 = "T10A,D20G,Q30E,A40T,P50L,G60R"
)

m <- mutstr(raw_mut_string, sep = ",")

m

names(m)

mstr(m)

mut(m)

m[1:2]

m[[2]]

intersect(m, m[1])

setdiff(m, m[1])

union(m, m[1])

}
