% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_bio.R
\name{nt2aa}
\alias{nt2aa}
\title{translate nucleotides into amino acids from the first character}
\usage{
nt2aa(x)
}
\arguments{
\item{x}{nucleotides characters}
}
\value{
amino acids characters
}
\description{
translate nucleotides into amino acids from the first character
}
\examples{
nt2aa(c("ATGAAA", "TTGCCC", "CTGTTT"))
}
