% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutations.R
\name{parse_aa_mut}
\alias{parse_aa_mut}
\title{parse substitutions and deletions of amino acids}
\usage{
parse_aa_mut(mut_vector)
}
\arguments{
\item{mut_vector}{mutation vector}
}
\value{
list of mutation table
}
\description{
parse substitutions and deletions of amino acids
}
\examples{
mut <- "G1:T10I,G2:D20N,G3:Q30E,G1:A40T,G3:P50L,G2:G60R"
parse_aa_mut(mut)

}
