% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bio.R
\name{parse_ANARCI_aaseq}
\alias{parse_ANARCI_aaseq}
\title{parse sequences from ANARCI vdj output}
\usage{
parse_ANARCI_aaseq(x, remove_gap = TRUE)
}
\arguments{
\item{x}{tibble from anarci_H.csv or anarci_KL.csv}

\item{remove_gap}{remove the gap caused by numbering, default as TRUE}
}
\value{
tibble
}
\description{
parse sequences from ANARCI vdj output
}
