% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bio.R
\name{parse_CellRanger_vdjseq}
\alias{parse_CellRanger_vdjseq}
\title{parse sequences from CellRanger vdj output}
\usage{
parse_CellRanger_vdjseq(x, file = NULL, fa_content = "sequence")
}
\arguments{
\item{x}{tibble from airr_rearrangement.tsv}

\item{file}{output file, default is NULL and will return a tibble}

\item{fa_content}{if the extension of file is 'fa|fasta', select a column to
write in the file. Can be one of \verb{sequence, seq_orf_nt, seq_orf_aa}}
}
\value{
tibble or NULL
}
\description{
parse sequences from CellRanger vdj output
}
