% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bio.R
\name{parse_IgBlast_shm}
\alias{parse_IgBlast_shm}
\title{parse somatic hypermutation from igblast output}
\usage{
parse_IgBlast_shm(igblast_fmt7)
}
\arguments{
\item{igblast_fmt7}{the filename of igblast result}
}
\value{
tibble
}
\description{
parse somatic hypermutation from igblast output
}
