% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bio.R
\name{parse_SingleR}
\alias{parse_SingleR}
\title{parse the output of SingleR}
\usage{
parse_SingleR(x)
}
\arguments{
\item{x}{DFrame object from SingleR::SingleR()}
}
\value{
tibble
}
\description{
parse the output of SingleR
}
