% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bio.R
\name{parse_vcf}
\alias{parse_vcf}
\title{parse vcf file with the help of reference genome and annotations.
It is still under development, which can not process more than 3 nt
substitutions in a single record row of vcf file, and can not process indels}
\usage{
parse_vcf(vcf, fa, gff)
}
\arguments{
\item{vcf}{\code{vcfR} object from \code{read_vcf}}

\item{fa}{\code{DNAStringSet} object from \code{read_fasta}}

\item{gff}{\code{gff} tibble from \code{read_gff}}
}
\value{
tibble
}
\description{
parse vcf file with the help of reference genome and annotations.
It is still under development, which can not process more than 3 nt
substitutions in a single record row of vcf file, and can not process indels
}
