% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{read_fasta}
\alias{read_fasta}
\title{read fasta file}
\usage{
read_fasta(x, type = "DNA")
}
\arguments{
\item{x}{file path}

\item{type}{one of \verb{DNA, RNA, AA}}
}
\value{
XXXStringSet object
}
\description{
read fasta file
}
