% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_bio.R
\name{reduction_SE}
\alias{reduction_SE}
\title{reduction from raw Seurat object created by read count matrix, including
normalization, variable features calling, scaling, PCA and UMAP}
\usage{
reduction_SE(x, use_dim = 30)
}
\arguments{
\item{x}{Seurat object}

\item{use_dim}{dims to use in UMAP reduction}
}
\value{
Seurat object
}
\description{
reduction from raw Seurat object created by read count matrix, including
normalization, variable features calling, scaling, PCA and UMAP
}
