% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortcut_bio.R
\name{SingleR_SE}
\alias{SingleR_SE}
\title{add SingleR celltype annotation for Seurat object}
\usage{
SingleR_SE(x, ref)
}
\arguments{
\item{x}{Seurat object}

\item{ref}{SummarizedExperiment object, the annotation reference of SingleR}
}
\value{
Seurat object
}
\description{
add SingleR celltype annotation for Seurat object
}
