% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{write_fasta}
\alias{write_fasta}
\title{write a sequence vector into fasta file}
\usage{
write_fasta(x, file)
}
\arguments{
\item{x}{sequence vector}

\item{file}{fasta file}
}
\description{
write a sequence vector into fasta file
}
